+++ packages/types/src/interfaces/treasury/definitions.ts
@@ -9,6 +9,37 @@ import { Definitions } from '../../types';
 export default {
   rpc: {},
   types: {
+    Bounty: {
+      proposer: 'AccountId',
+      value: 'Balance',
+      fee: 'Balance',
+      curatorDeposit: 'Balance',
+      bond: 'Balance',
+      status: 'BountyStatus'
+    },
+    BountyIndex: 'u32',
+    BountyStatus: {
+      _enum: {
+        Proposed: 'Null',
+        Approved: 'Null',
+        Funded: 'Null',
+        CuratorProposed: 'BountyStatusCuratorProposed',
+        Active: 'BountyStatusActive',
+        PendingPayout: 'BountyStatusPendingPayout'
+      }
+    },
+    BountyStatusActive: {
+      curator: 'AccountId',
+      updateDue: 'BlockNumber'
+    },
+    BountyStatusCuratorProposed: {
+      curator: 'AccountId'
+    },
+    BountyStatusPendingPayout: {
+      curator: 'AccountId',
+      beneficiary: 'AccountId',
+      unlockAt: 'BlockNumber'
+    },
     OpenTip: {
       reason: 'Hash',
       who: 'AccountId',